// Financials/ClientApp/src/app/interfaces/currentbullbearrun.ts
export class CurrentBullBearRun {
  id: number;
  securityId: number;
  runType: number;

  dateCreated: Date;
  dateModified: Date;
  runStartDate: Date;
  runEndDate: Date;

  highDate: Date;
  lowDate: Date;

  startRunPrice: number;
  endRunPrice: number;
  lowPrice: number;
  highPrice: number;
  percentRangeCheck: number;

}